Fix to allow navigation for about:blank
Running an app with a google maps iframe throws `ERROR Internal navigation rejected - <allow-navigation> not set for url='about:blank'` in the XCode console. This caused google maps embeds to only load partially on iOS. Using a wildcard in the rule fixes this and the error is no longer logged.